Move node pool utils to a common place
