+++ src/components/MAPI/organizations/OrganizationDetailGeneral/utils.ts
@@ -2,6 +2,13 @@ import ErrorReporter from 'lib/errors/ErrorReporter';
 import { HttpClientFactory } from 'lib/hooks/useHttpClientFactory';
 import { IOAuth2Provider } from 'lib/OAuth2/OAuth2';
 import { compare } from 'lib/semver';
+import {
+  fetchMasterListForCluster,
+  fetchNodePoolListForCluster,
+  fetchProviderNodePoolsForNodePools,
+  getMachineTypes,
+  IMachineType,
+} from 'MAPI/utils';
 import * as applicationv1alpha1 from 'model/services/mapi/applicationv1alpha1';
 import * as capiv1alpha3 from 'model/services/mapi/capiv1alpha3';
 import * as capiexpv1alpha3 from 'model/services/mapi/capiv1alpha3/exp';
@@ -95,33 +102,6 @@ function fetchSingleClusterSummaryKey(cluster: capiv1alpha3.ICluster): string {
   return `fetchSingleClusterSummary/${cluster.metadata.namespace}/${cluster.metadata.name}`;
 }
 
-async function fetchMasterListForCluster(
-  httpClientFactory: HttpClientFactory,
-  auth: IOAuth2Provider,
-  cluster: capiv1alpha3.ICluster
-) {
-  const infrastructureRef = cluster.spec?.infrastructureRef;
-  if (!infrastructureRef) {
-    return Promise.reject(
-      new Error('There is no infrastructure reference defined.')
-    );
-  }
-
-  switch (infrastructureRef.kind) {
-    case capzv1alpha3.AzureCluster:
-      return capzv1alpha3.getAzureMachineList(httpClientFactory(), auth, {
-        labelSelector: {
-          matchingLabels: {
-            [capiv1alpha3.labelCluster]: cluster.metadata.name,
-          },
-        },
-      });
-
-    default:
-      return Promise.reject(new Error('Unsupported provider.'));
-  }
-}
-
 function appendMasterStats(
   masters: capzv1alpha3.IAzureMachine[],
   machineTypes: Record<string, IMachineType>,
@@ -147,43 +127,6 @@ function appendMasterStats(
   }
 }
 
-async function fetchNodePoolListForCluster(
-  httpClientFactory: HttpClientFactory,
-  auth: IOAuth2Provider,
-  cluster: capiv1alpha3.ICluster
-) {
-  const infrastructureRef = cluster.spec?.infrastructureRef;
-  if (!infrastructureRef) {
-    return Promise.reject(
-      new Error('There is no infrastructure reference defined.')
-    );
-  }
-
-  switch (infrastructureRef.kind) {
-    case capzv1alpha3.AzureCluster:
-      return capiexpv1alpha3.getMachinePoolList(httpClientFactory(), auth, {
-        labelSelector: {
-          matchingLabels: {
-            [capiv1alpha3.labelCluster]: cluster.metadata.name,
-          },
-        },
-      });
-
-    // TODO(axbarsan): Use CAPA type once available.
-    case 'AWSCluster':
-      return capiv1alpha3.getMachineDeploymentList(httpClientFactory(), auth, {
-        labelSelector: {
-          matchingLabels: {
-            [capiv1alpha3.labelCluster]: cluster.metadata.name,
-          },
-        },
-      });
-
-    default:
-      return Promise.reject(new Error('Unsupported provider.'));
-  }
-}
-
 function appendNodePoolsStats(
   nodePools: capiv1alpha3.IMachineDeployment[] | capiexpv1alpha3.IMachinePool[],
   summary: ui.IOrganizationDetailClustersSummary
@@ -194,38 +137,6 @@ function appendNodePoolsStats(
       summary.workerNodesCount += nodePool.status.readyReplicas;
     }
   }
-}
-
-async function fetchProviderNodePoolsForNodePools(
-  httpClientFactory: HttpClientFactory,
-  auth: IOAuth2Provider,
-  nodePools: capiv1alpha3.IMachineDeployment[] | capiexpv1alpha3.IMachinePool[]
-) {
-  return Promise.all(
-    nodePools.map(
-      (np: capiv1alpha3.IMachineDeployment | capiexpv1alpha3.IMachinePool) => {
-        const infrastructureRef = np.spec?.template.spec.infrastructureRef;
-        if (!infrastructureRef) {
-          return Promise.reject(
-            new Error('There is no infrastructure reference defined.')
-          );
-        }
-
-        switch (infrastructureRef.kind) {
-          case capzexpv1alpha3.AzureMachinePool:
-            return capzexpv1alpha3.getAzureMachinePool(
-              httpClientFactory(),
-              auth,
-              np.metadata.namespace!,
-              infrastructureRef.name
-            );
-
-          default:
-            return Promise.reject(new Error('Unsupported provider.'));
-        }
-      }
-    )
-  );
 }
 
 function appendProviderNodePoolsStats(
@@ -251,44 +162,6 @@ function appendProviderNodePoolsStats(
       summary.workerNodesMemory += machineTypeProperties.memory * readyReplicas;
     }
   }
-}
-
-interface IMachineType {
-  cpu: number;
-  memory: number;
-}
-
-function getMachineTypes(): Record<string, IMachineType> {
-  const machineTypes: Record<string, IMachineType> = {};
-
-  if (window.config.awsCapabilitiesJSON) {
-    const rawCapabilities: Record<string, IRawAWSInstanceType> = JSON.parse(
-      window.config.awsCapabilitiesJSON
-    );
-
-    for (const [name, properties] of Object.entries(rawCapabilities)) {
-      machineTypes[name] = {
-        cpu: properties.cpu_cores,
-        memory: properties.memory_size_gb,
-      };
-    }
-  }
-
-  if (window.config.azureCapabilitiesJSON) {
-    const rawCapabilities: Record<string, IRawAzureInstanceType> = JSON.parse(
-      window.config.azureCapabilitiesJSON
-    );
-
-    for (const [name, properties] of Object.entries(rawCapabilities)) {
-      machineTypes[name] = {
-        cpu: properties.numberOfCores,
-        // eslint-disable-next-line no-magic-numbers
-        memory: properties.memoryInMb / 1000,
-      };
-    }
-  }
-
-  return machineTypes;
 }
 
 function mergeClusterSummaries(
